+++ src/__test__/utils/nodesEdgesUtils/removeEdge.test.js
@@ -1,5 +1,5 @@
 import { DataSet } from 'vis-data'
-import { OPERATION_TYPE_ADD, OPERATION_TYPE_TOGGLE } from '../../../constants/store'
+import { OPERATION_TYPE_ADD, OPERATION_TYPE_ARRAY_DELETE_INDEX, OPERATION_TYPE_TOGGLE } from '../../../constants/store'
 import store from '../../../store'
 import removeEdge from '../../../utils/nodesEdgesUtils/removeEdge'
 
@@ -9,6 +9,11 @@ const edge = {
   from: '1',
   to: '3'
 }
+const dataTableTriples = [{
+  edge: '12',
+  from: '1',
+  to: '3'
+}]
 
 describe('removeEdge', () => {
   afterEach(() => {
@@ -18,7 +23,8 @@ describe('removeEdge', () => {
   it('should not remove edge if not existing', async () => {
     const availableEdges = new DataSet()
     store.getState = () => ({
-      availableEdges
+      availableEdges,
+      dataTableTriples
     })
 
     await removeEdge({
@@ -37,7 +43,8 @@ describe('removeEdge', () => {
     ])
 
     store.getState = () => ({
-      availableEdges
+      availableEdges,
+      dataTableTriples
     })
 
     await removeEdge({
@@ -49,6 +56,20 @@ describe('removeEdge', () => {
 
     expect(updateStoreValue.mock.calls).toEqual([
       [['availableEdgesCount'], OPERATION_TYPE_ADD, -1],
+      [
+        [
+          'dataTableTriples',
+        ],
+        OPERATION_TYPE_ARRAY_DELETE_INDEX,
+        0,
+      ],
+      [
+        [
+          'dataTableTriplesWithLabels',
+        ],
+        OPERATION_TYPE_ARRAY_DELETE_INDEX,
+        0,
+      ],
       [['nodesEdges', edge.from], OPERATION_TYPE_TOGGLE, edge.id],
       [['nodesEdges', edge.to], OPERATION_TYPE_TOGGLE, edge.id],
     ])