+++ src/utils/nodesEdgesUtils/removeEdge.js
@@ -1,4 +1,4 @@
-import { OPERATION_TYPE_TOGGLE, OPERATION_TYPE_ADD } from '../../constants/store'
+import { OPERATION_TYPE_TOGGLE, OPERATION_TYPE_ADD, OPERATION_TYPE_ARRAY_DELETE_INDEX } from '../../constants/store'
 import store from '../../store'
 import getEdge from './getEdge'
 
@@ -13,22 +13,31 @@ const removeEdge = ({
   updateStoreValue
 }) => {
   const {
-    availableEdges
+    availableEdges,
+    dataTableTriples
   } = store.getState()
-
-  const isVisible = getEdge(edge.id)
-
-  if (isVisible === null) return false
-
-  availableEdges.remove(edge.id)
-
-  updateStoreValue(['availableEdgesCount'], OPERATION_TYPE_ADD, -1)
 
   const {
     from,
     to,
     id
   } = edge
+
+  const isVisible = getEdge(id)
+
+  if (isVisible === null) return false
+
+  availableEdges.remove(id)
+
+  updateStoreValue(['availableEdgesCount'], OPERATION_TYPE_ADD, -1)
+
+  // remove from datatable rows
+  if (dataTableTriples.length > 0) {
+    const edgeIndex = dataTableTriples.findIndex((triple) => triple.edge === id)
+
+    updateStoreValue(['dataTableTriples'], OPERATION_TYPE_ARRAY_DELETE_INDEX, edgeIndex)
+    updateStoreValue(['dataTableTriplesWithLabels'], OPERATION_TYPE_ARRAY_DELETE_INDEX, edgeIndex)
+  }
 
   updateStoreValue(['nodesEdges', from], OPERATION_TYPE_TOGGLE, id)
   updateStoreValue(['nodesEdges', to], OPERATION_TYPE_TOGGLE, id)